rk3399: Support upgrade_available env
upgrade_available env is used to preserve the status
for full partition upgrade systems.

By default, it is set 0 assuming upgrade won't need.

But for full upgrade partitions, it is set to 1 so-that
the system will reboot from recovery and do upgrade for
default partitions.

Once the full upgrade is done, it ise set to 2 and reboot
for main default partition.

Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>
Change-Id: Iee4a5ad2f551b33ac92fb72bf1480fe67e1ad8e5