rock960: Update the partition layout
We have redesigned update system to have 3 key partitions,
the main system partition is rootfs, the golden or recovery
partition is rootfs-rollback and the full image update via
updater partition.

So, move the rootfs to end of the partition layout, so-that
the static sized partition before that would allocate and
give rest of the space for rootfs.

This is default layout for 16GiB eMMC and the same can
reuse it for 8MiB eMMC by grabbing some space from data
partition.

16GiB eMMC layout:
Size         Name
3.9 MiB      loader1
4.0 MiB      loader2
4.0 MiB      trust
112.0 MiB    boot
112.0 MiB    boot-rollback
64.0 MiB     updater
1.3 GiB      rootfs-rollback
1024.0 MiB   swap
3.0 GiB      data
9.0 GiB      rootfs

Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>
Change-Id: Ia92fcb1e4333821bb9662a5aac54d08042c536d6